refactor(flight): split tail-lift / phase-7 altitude statics into ParsekFlight.TailLift.cs
